+++ src/shared/lib/index.ts
@@ -1,3 +1,3 @@
 export { devLogger } from "./devLogger";
 export { useCurrentCalendarId } from "./useCurrentCalendarId";
-export { cn, getCategoryColorClass } from "./utils";
+export { cn } from "./utils";
